Align pools page and staking page styling
