+++ archetypes/Pools/FilterSelects/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import BaseFilters from '~/components/BaseFilters';
 import { LogoTicker } from '~/components/General';
 import TWButtonGroup from '~/components/General/TWButtonGroup';
 import * as Styles from './styles';
@@ -76,21 +77,21 @@ const FilterSelects: React.FC<FilterSelectsProps> = ({ state, dispatch }) => {
         dispatch({ type: 'setRebalanceFocus', focus: option as RebalanceEnum });
 
     return (
-        <Styles.Container>
-            <Styles.SearchInput placeholder="Search" value={state.search} onChange={onSearchInputChange} />
-            <Styles.FilterPopup
+        <BaseFilters.Container>
+            <BaseFilters.SearchInput placeholder="Search" value={state.search} onChange={onSearchInputChange} />
+            <BaseFilters.FilterPopup
                 preview={
-                    <Styles.Preview>
-                        <Styles.FilterIcon />
+                    <BaseFilters.Preview>
+                        <BaseFilters.FilterIcon />
                         Filter Results
-                    </Styles.Preview>
+                    </BaseFilters.Preview>
                 }
                 buttonClasses="action-button"
             >
-                <Styles.Content>
+                <BaseFilters.Content>
                     <div>
-                        <Styles.Heading>Market</Styles.Heading>
-                        <Styles.Dropdown
+                        <BaseFilters.Heading>Market</BaseFilters.Heading>
+                        <BaseFilters.Dropdown
                             variant="default"
                             iconSize="xs"
                             placeHolderIcon={
@@ -103,25 +104,25 @@ const FilterSelects: React.FC<FilterSelectsProps> = ({ state, dispatch }) => {
                             onSelect={onMarketSelect}
                         />
                     </div>
-                    <Styles.Wrapper>
-                        <Styles.DropdownContainer>
-                            <Styles.Text>Collateral</Styles.Text>
-                            <Styles.Dropdown
+                    <BaseFilters.Wrapper>
+                        <BaseFilters.DropdownContainer>
+                            <BaseFilters.Text>Collateral</BaseFilters.Text>
+                            <BaseFilters.Dropdown
                                 value={state.collateralFilter ?? 'All'}
                                 options={COLLATERAL_FILTER_OPTIONS}
                                 onSelect={onCollateralFilterSelect}
                             />
-                        </Styles.DropdownContainer>
-                        <Styles.DropdownContainer>
-                            <Styles.Text>Power Leverage</Styles.Text>
-                            <Styles.Dropdown
+                        </BaseFilters.DropdownContainer>
+                        <BaseFilters.DropdownContainer>
+                            <BaseFilters.Text>Power Leverage</BaseFilters.Text>
+                            <BaseFilters.Dropdown
                                 value={state.leverageFilter}
                                 options={LEVERAGE_FILTER_OPTIONS}
                                 onSelect={onLeverageFilterSelect}
                             />
-                        </Styles.DropdownContainer>
-                    </Styles.Wrapper>
-                    <Styles.DropdownContainer>
+                        </BaseFilters.DropdownContainer>
+                    </BaseFilters.Wrapper>
+                    <BaseFilters.DropdownContainer>
                         <TWButtonGroup
                             size="responsive"
                             value={state.deltaDenotation}
@@ -132,8 +133,8 @@ const FilterSelects: React.FC<FilterSelectsProps> = ({ state, dispatch }) => {
                             options={DENOTATION_OPTIONS}
                             fullWidth
                         />
-                    </Styles.DropdownContainer>
-                    <Styles.DropdownContainer>
+                    </BaseFilters.DropdownContainer>
+                    <BaseFilters.DropdownContainer>
                         <TWButtonGroup
                             size="responsive"
                             value={state.rebalanceFocus}
@@ -142,10 +143,10 @@ const FilterSelects: React.FC<FilterSelectsProps> = ({ state, dispatch }) => {
                             options={REBALANCE_OPTIONS}
                             fullWidth
                         />
-                    </Styles.DropdownContainer>
-                </Styles.Content>
-            </Styles.FilterPopup>
-        </Styles.Container>
+                    </BaseFilters.DropdownContainer>
+                </BaseFilters.Content>
+            </BaseFilters.FilterPopup>
+        </BaseFilters.Container>
     );
 };
