+++ archetypes/Pools/styles.tsx
@@ -1,47 +1,6 @@
 import styled from 'styled-components';
 import { default as UnstyledButton } from '~/components/General/Button';
-import { Container as UnstyledContainer } from '~/components/General/Container';
 import { default as UnstyledLoading } from '~/components/General/Loading';
-
-export const Header = styled.div`
-    margin-bottom: 2rem;
-
-    @media ${({ theme }) => theme.device.lg} {
-        display: flex;
-        gap: 20px;
-    }
-`;
-
-export const Heading = styled.h1`
-    color: ${({ theme }) => theme.fontColor.primary};
-    margin: 2rem 0 0.5rem;
-    font-size: 1.875rem;
-    line-height: 2.25rem;
-    font-weight: 600;
-`;
-
-export const SubHeading = styled.div`
-    margin-bottom: 1.5rem;
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    font-weight: 300;
-
-    @media ${({ theme }) => theme.device.lg} {
-        margin-bottom: 0;
-    }
-`;
-
-export const Link = styled.a.attrs({
-    target: '_blank',
-    rel: 'noreferrer',
-})`
-    text-decoration: underline;
-    color: #3da8f5;
-`;
-
-export const Container = styled(UnstyledContainer)`
-    margin-bottom: 2.5rem;
-`;
 
 export const Loading = styled(UnstyledLoading)`
     margin: 2.5rem auto 0;
